Fix NSLog string formatting again
The issue is that NSUInteger requires a printf specifier of "%u" on 32bit, but "%lu" on 64bit. Apple's recommended solution is to cast the parameter to (unsigned long) and use "%lu".

+++ Couch/CouchDatabase.m
@@ -142,14 +142,14 @@ static const NSUInteger kDocRetainLimit = 50;
     if (!_busyDocuments)
         _busyDocuments = [[NSCountedSet alloc] init];
     [_busyDocuments addObject: resource];
-    NSLog(@">>>>>> %lu docs being updated", _busyDocuments.count);
+    NSLog(@">>>>>> %lu docs being updated", (unsigned long)_busyDocuments.count);
 }
 
 
 - (void) endDocumentOperation: (CouchResource*)resource {
     NSAssert([_busyDocuments containsObject: resource], @"unbalanced endDocumentOperation call: %p %@", resource, resource);
     [_busyDocuments removeObject: resource];
-    NSLog(@"<<<<<< %lu docs being updated", _busyDocuments.count);
+    NSLog(@"<<<<<< %lu docs being updated", (unsigned long)_busyDocuments.count);
     if (_busyDocuments.count == 0)
         [self processDeferredChanges];
 }
